Reorganize CommitFilters structure
