+++ src/main/services/MiningModule.groovy
@@ -6,6 +6,9 @@ import interfaces.CommitFilter
 import interfaces.DataCollector
 import interfaces.OutputProcessor
 import interfaces.ProjectProcessor
+
+import services.commitFilters.InCommitListAndHasMutuallyModifiedMethodsFilter
+
 import services.modifiedLinesCollector.ModifiedLinesCollector
 import services.projectProcessors.FilterNonExistentProjectsProcessor
 import services.projectProcessors.ForkAndEnableTravisProcessor
@@ -22,7 +25,7 @@ public class MiningModule extends AbstractModule {
         dataCollectorBinder.addBinding().to(StatisticsCollectorImpl.class)
         dataCollectorBinder.addBinding().to(BuildRequester.class)
 
-        bind(CommitFilter.class).to(CommitFilterImpl.class)
+        bind(CommitFilter.class).to(InCommitListAndHasMutuallyModifiedMethodsFilter.class)
 
         projectProcessorBinder.addBinding().to(FilterNonExistentProjectsProcessor.class)
         projectProcessorBinder.addBinding().to(ForkAndEnableTravisProcessor.class)
